Do not set web hooks if Host is localhost
Docker does not use the default localhost IP address. One way to get around this it to check if the Host field is set to localhost.

If the host is not using the default 80 port, the used port number will be set in, and part of, the Host field which is why we check for the substring localhost.