Fix payment form not appearing on block-based themes
The localization relies on the `nets-easy-checkout` script. However, the current action on which it is enqueued, `wc_dibs_before_checkout_form`, occurs too early in the process, specifically during the templating stage. At this point, the necessary `dibs_load_js` action has not yet been triggered. Since we depend on the localized variables provided by `dibs_load_js`, the checkout fails to initialize properly.

It appears that WooCommerce has altered the order in which these actions are triggered. In block-based themes, the template actions occur before the scripts are enqueued. However, in classical themes, the order is reversed, with the scripts being enqueued before the template actions.