Do not update merchant order reference for order pay
Currently, we retrieve the `$checkout_flow` directly from the settings, however, for order pay, we use the redirect flow. Thus the `$checkout_flow` is incorrectly set to 'embedded' which results in order reference update request which should not be issued for order pay.

By the time, we reach `wc_dibs_confirm_dibs_order`, we're no longer on the order-pay endpoint, and can no longer determine if it was an order pay purchase. Therefore, we have to save the checkout flow in the order meta data.